refactor(conformance): move the shape benchmark to a criterion bench
Replaces the throwaway `examples/map_bench.rs` with
`benches/map_bench.rs`, so the numbers behind the view-vs-owned claims come
from criterion's sampling rather than a hand-rolled `Instant` loop. The
earlier timings were noisy enough to read as flat in `n` where the real
curve narrows as the fixed index build amortizes.

The crate grows a `[lib]` target. It was bin-only, so a bench could not
reach the generated types without re-`include!`-ing `_include.rs` and
duplicating the module scaffolding from `main.rs`. `generated`, `registry`
and the harness-violation conversions move to `src/lib.rs`; `main.rs` keeps
the stdin/stdout executor and `run_case`. `registry::dispatch` and
`CaseOutcome` become `pub` for the binary to reach.

criterion is a dev-dependency with default features off, so the plotting
and gnuplot trees stay out; `cargo_bench_support` is the only feature
needed. `harness = false` means `cargo test --all-targets` runs each
benchmark once, which keeps it from rotting without lengthening CI
noticeably — the full workspace test run is 30s.

First benchmark in the repo, so there was no existing bench convention to
follow. Verified with `cargo bench -p protovalidate-buffa-conformance`;
fmt, clippy, the workspace tests and the upstream harness (2872/2872 in
both default and --strict_error) all still pass.
